Revert "Wait for changes in Daemon executor as well"
We now fixed the problems in FileTimeStampInspector, so we
don't need to wait any more.

This reverts commit d75762a4
